step3, package 정리
