+++ src/main/kotlin/racing/controller/RacingControllerImpl.kt
@@ -1,14 +1,14 @@
-package step3.racing.controller
+package racing.controller
 
-import step3.racing.model.car.Car
-import step3.racing.model.generator.NumberGenerator
-import step3.racing.view.input.InputView
-import step3.racing.view.result.ResultView
+import racing.model.car.Car
+import racing.model.generator.NumberGenerator
+import racing.view.input.InputView
+import racing.view.result.ResultView
 
 internal class RacingControllerImpl(
     private val inputView: InputView,
     private val resultView: ResultView,
-    private val numberGenerator: NumberGenerator,
+    private val numberGenerator: racing.model.generator.NumberGenerator,
 ) : RacingController {
     override fun start() {
         val carCount = inputView.promptAndValidateCarCountInput()
@@ -24,17 +24,17 @@ internal class RacingControllerImpl(
         }
     }
 
-    private fun getCarList(carCount: Int): List<Car> {
+    private fun getCarList(carCount: Int): List<racing.model.car.Car> {
         return List(carCount) {
-            Car(
-                forwardLimit = Car.DEFAULT_FORWARD_LIMIT,
+            racing.model.car.Car(
+                forwardLimit = racing.model.car.Car.DEFAULT_FORWARD_LIMIT,
                 numberGenerator = numberGenerator,
             )
         }
     }
 
-    private fun startRound(cars: List<Car>) {
-        fun move(car: Car) = car.move()
+    private fun startRound(cars: List<racing.model.car.Car>) {
+        fun move(car: racing.model.car.Car) = car.move()
         cars.forEach(::move)
     }
 }
