+++ src/main/kotlin/racing/model/car/Car.kt
@@ -1,14 +1,14 @@
-package step3.racing.model.car
+package racing.model.car
 
-import step3.racing.model.generator.NumberGenerator
-import step3.racing.model.generator.RandomGeneratorFactory
+import racing.model.generator.NumberGenerator
+import racing.model.generator.RandomGeneratorFactory
 
 class Car(
-    var carCount: Int = DEFAULT_CAR_COUNT,
-    var attemptCount: Int = DEFAULT_ATTEMPT_COUNT,
-    position: Int = DEFAULT_POSITION,
-    private val forwardLimit: Int = DEFAULT_FORWARD_LIMIT,
-    private val numberGenerator: NumberGenerator = RandomGeneratorFactory.newInstance(),
+    var carCount: Int = racing.model.car.Car.Companion.DEFAULT_CAR_COUNT,
+    var attemptCount: Int = racing.model.car.Car.Companion.DEFAULT_ATTEMPT_COUNT,
+    position: Int = racing.model.car.Car.Companion.DEFAULT_POSITION,
+    private val forwardLimit: Int = racing.model.car.Car.Companion.DEFAULT_FORWARD_LIMIT,
+    private val numberGenerator: racing.model.generator.NumberGenerator = RandomGeneratorFactory.newInstance(),
 ) {
     var position: Int = position
         private set
