+++ src/main/kotlin/racing/model/generator/RandomGeneratorFactory.kt
@@ -1,0 +1,9 @@
+package racing.model.generator
+
+object RandomGeneratorFactory {
+    private val DEFAULT_RANDOM_RANGE = 0..9
+
+    fun newInstance(range: IntRange = DEFAULT_RANDOM_RANGE): racing.model.generator.NumberGenerator {
+        return RandomGeneratorImpl(range)
+    }
+}
