+++ src/main/kotlin/racing/model/generator/RandomGeneratorImpl.kt
@@ -1,0 +1,7 @@
+package racing.model.generator
+
+internal class RandomGeneratorImpl(private val range: IntRange) : racing.model.generator.NumberGenerator {
+    override fun generator(): Int {
+        return range.random()
+    }
+}
