+++ src/main/kotlin/racing/view/input/ConsoleInputViewFactory.kt
@@ -1,4 +1,4 @@
-package step3.racing.view.input
+package racing.view.input
 
 object ConsoleInputViewFactory {
     fun newInstance(): InputView {
