+++ src/main/kotlin/racing/view/input/InputView.kt
@@ -1,4 +1,4 @@
-package step3.racing.view.input
+package racing.view.input
 
 sealed interface InputView {
     fun printCarCountQuestion()
