+++ src/main/kotlin/racing/view/result/ResultView.kt
@@ -1,0 +1,9 @@
+package racing.view.result
+
+import racing.model.car.Car
+
+sealed interface ResultView {
+    fun printOutputTitle()
+
+    fun displayCarMovement(cars: List<racing.model.car.Car>)
+}
