--- src/main/kotlin/step3/racing/model/generator/RandomGeneratorFactory.kt
@@ -1,9 +1,0 @@
-package step3.racing.model.generator
-
-object RandomGeneratorFactory {
-    private val DEFAULT_RANDOM_RANGE = 0..9
-
-    fun newInstance(range: IntRange = DEFAULT_RANDOM_RANGE): NumberGenerator {
-        return RandomGeneratorImpl(range)
-    }
-}
