--- src/main/kotlin/step3/racing/model/generator/RandomGeneratorImpl.kt
@@ -1,7 +1,0 @@
-package step3.racing.model.generator
-
-internal class RandomGeneratorImpl(private val range: IntRange) : NumberGenerator {
-    override fun generator(): Int {
-        return range.random()
-    }
-}
