--- src/main/kotlin/step3/racing/view/result/ResultView.kt
@@ -1,9 +1,0 @@
-package step3.racing.view.result
-
-import step3.racing.model.car.Car
-
-sealed interface ResultView {
-    fun printOutputTitle()
-
-    fun displayCarMovement(cars: List<Car>)
-}
