+++ src/test/kotlin/study/racing/model/car/CarTest.kt
@@ -1,18 +1,17 @@
-package study.racing
+package study.racing.model.car
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
-import step3.racing.model.car.Car
-import step3.racing.model.generator.RandomGeneratorFactory
+import racing.model.generator.RandomGeneratorFactory
 
 class CarTest {
     @Test
     fun `주어진 횟수 동안 n대의 자동차는 4미만인 경우 멈춘다`() {
         val numberGenerator = RandomGeneratorFactory.newInstance(0..3)
-        val car = Car(numberGenerator = numberGenerator)
+        val car = racing.model.car.Car(numberGenerator = numberGenerator)
         car.move()
         car.position shouldBe 0
     }
@@ -20,7 +19,7 @@ class CarTest {
     @Test
     fun `주어진 횟수 동안 n대의 자동차는 4이상인 경우 전진한다`() {
         val numberGenerator = RandomGeneratorFactory.newInstance(4..9)
-        val car = Car(numberGenerator = numberGenerator)
+        val car = racing.model.car.Car(numberGenerator = numberGenerator)
         car.move()
         car.position shouldBe 1
     }
@@ -35,7 +34,7 @@ class CarTest {
             RandomGeneratorFactory.newInstance(
                 randomMin..randomMax,
             )
-        val car = Car(numberGenerator = numberGenerator)
+        val car = racing.model.car.Car(numberGenerator = numberGenerator)
         car.move()
         car.position shouldBe 0
     }
@@ -47,7 +46,7 @@ class CarTest {
         randomMax: Int,
     ) {
         val numberGenerator = RandomGeneratorFactory.newInstance(randomMin..randomMax)
-        val car = Car(numberGenerator = numberGenerator)
+        val car = racing.model.car.Car(numberGenerator = numberGenerator)
         car.move()
         car.position shouldBe 1
     }
@@ -58,7 +57,7 @@ class CarTest {
         carCount: Int,
         attemptCount: Int,
     ) {
-        val car = Car()
+        val car = racing.model.car.Car()
         car.carCount = carCount
         car.carCount shouldBe carCount
         car.attemptCount = attemptCount
@@ -67,7 +66,7 @@ class CarTest {
 
     @Test
     fun `자동차 대수는 최소 입력 테스트`() {
-        val car = Car()
+        val car = racing.model.car.Car()
         shouldThrow<IllegalArgumentException> {
             car.carCount = 0
             car.move()
